Conflicts:
	www/sony_o2o/__init__.py

+++ www/sony_o2o/views/guest/Guest.py
@@ -4,36 +4,38 @@ from flask.ext import restful
 from flask import request
 from sony_o2o.libs import auth
 from o2olib import GuestService
+from JsonResult import JsonResult as JR
 
 
 class Guest(restful.Resource):
 
     #@auth.require_login()
     def get(self, id):
-        guest = {"id": id}
-        return guest
+        guests = GuestService.get_guests()
+        if len(guests) == 0:
+            rs = {}
+        else:
+            rs = guests[0]
+        return JR(rs).to_dic()
 
     def post(self):
-        guest = {"name": 'luyan'}
-        GuestService.add_guest(guest)
-        return guest
+        guest = dict(request.form)
+        rs = GuestService.add_guest(guest)
+        return JR(rs).to_dic()
 
-    def put(self, id):
-        guest = {}
-        guest['id'] = id
-        guest['name'] = request.form.get('name')
-        GuestService.update_guest(guest)
-        return guest
+    def put(self):
+        guest = dict(request.form)
+        rs = GuestService.update_guest(guest)
+        return JR(rs).to_dic()
 
-    def delete(self):
-        GuestService.delete_guest(id)
-        return id
-
+    def delete(self,id):
+        if id:
+           rs = GuestService.delete_guest(id)
+        return JR(rs).to_dic()
 
 class GuestList(restful.Resource):
 
     #@auth.require_login()
     def get(self):
-        print 'you search me'
-        print request.args.get('name')
-        return ''
+        con_dic = dict(request.args)
+        return GuestService.get_guests(con_dic)
